+++ src/main.py
@@ -1,5 +1,4 @@
 import gamemanager
 
 if __name__ == "__main__":
-    game = gamemanager.GameManager()
-    game.initialize()
+    gamemanager.GameManager()