+++ src/utils/stage.py
@@ -1,5 +1,5 @@
-Height = 800
-Width = 1000
+HEIGHT = 800
+WIDTH = 1000
 
 class Stage:
     stage = None