+++ src/utils/time.py
@@ -1,4 +1,4 @@
-from pygame.time import Clock, get_ticks
+from pygame.time import Clock
 
 class Time:
     dt = 0
@@ -7,4 +7,4 @@ class Time:
     @classmethod
     def update(cls, clock: Clock):
         cls.dt = clock.get_time() / 1000
-        cls.time += get_ticks() / 1000
+        cls.time += Clock.get_ticks() / 1000