+++ dokumentaatio/changelog.md
@@ -3,3 +3,8 @@
 - a basic system to run the game
 - physics for circles
 - basically the entirety of the player controller
+
+## Week 4
+**Added:**
+- polygons the player can stick to
+- collision between circle and polygon colliders